refactor: migrate styles from css to scss
